% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_time.r
\name{align_time}
\alias{align_time}
\title{Align two time series (optionally with multiple channels)}
\usage{
align_time(
  m,
  plot = FALSE,
  verbose = FALSE,
  max_iter = 20,
  get_lwr_bound = TRUE,
  max_combn = 50000,
  shifts = 1:(dim(m)[2]) - round((dim(m)[2])/2),
  warn_max_iter = TRUE
)
}
\arguments{
\item{m}{should be a n x T x K array, where n is the sample size, T is the period length, and K is the number of channels. A n x T matrix can also be used.}
}
\description{
Align two time series (optionally with multiple channels)
}
