% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTemp.R
\name{extractTemp}
\alias{extractTemp}
\title{Extract temperature data from an accelerometer file}
\usage{
extractTemp(x)
}
\arguments{
\item{x}{Location of accelerometer csv file}
}
\description{
This function reads in an accelerometer file and extracts temperature data. 
Returns a dataframe with two columns: a time stamp, and the temp record
}
\examples{
extractTemp()
}
