% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDomPeriod.R
\name{findDomPeriod}
\alias{findDomPeriod}
\title{Find the dominant period and weight of a single accelerometer axis}
\usage{
findDomPeriod(x)
}
\arguments{
\item{x}{A single acceleration axis}
}
\description{
This function calculates the dominant period of an accelerometer axis using an autoregressive modelling approach. It also returns a weight ((dominant - 2nd dominant)/dominant).
Returns two values: the dominant period, and a weight
}
\examples{
findDomPeriod()
}
