% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_processing.r
\name{fit_01_wave}
\alias{fit_01_wave}
\title{Fit a cosine wave}
\usage{
fit_01_wave(y, x, plot_it = FALSE, ...)
}
\arguments{
\item{y}{vector representing a value that changes over time}

\item{x}{time variable. must be in units of cycles (x ranging from 0 to 1 denotes 1 cycle). multiple cycles should be treated as "overlaid" on the same 0-1 interval.}

\item{plot_it}{If TRUE, a plot of the fit is shown}
}
\description{
Fit a cosine wave
}
