% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_processing.r
\name{fit_depmix_individual}
\alias{fit_depmix_individual}
\title{Fitting an HMM to an individual (shortcut function)}
\usage{
fit_depmix_individual(
  y,
  nstates = 3,
  family = gaussian(),
  ntimes = length(y),
  ...
)
}
\arguments{
\item{y}{activity; encouraged to be log(activity + 1)}
}
\description{
Fitting an HMM to an individual (shortcut function)
}
