% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_processing.r
\name{fit_fast_minmax_window}
\alias{fit_fast_minmax_window}
\title{Get M10, L5 or other window summaries, in linear time
For L5 M10 - the smallest, or largest, window of size W, where W is in difftime units.
This faster version tracks a moving window, rather than summing over all elements in the window each time. Additional conserations are given to adjust for NAs.}
\usage{
fit_fast_minmax_window(
  x,
  posix,
  minmax,
  time_length = 1 * hr1,
  return_type = "value",
  nonzero_min_proportion = 0.1,
  max_gap = 2 * hr1
)
}
\description{
Get M10, L5 or other window summaries, in linear time
For L5 M10 - the smallest, or largest, window of size W, where W is in difftime units.
This faster version tracks a moving window, rather than summing over all elements in the window each time. Additional conserations are given to adjust for NAs.
}
