% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_processing.r
\name{get_viterbi_with_columns_sorted_by_state}
\alias{get_viterbi_with_columns_sorted_by_state}
\title{Get object from HMM model to plot}
\usage{
get_viterbi_with_columns_sorted_by_state(
  mod,
  y,
  include_mean_in_level_name = FALSE
)
}
\description{
Get object from HMM model to plot
}
