% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_processing.r
\name{group_by_shifted_day}
\alias{group_by_shifted_day}
\title{Shortcut function Aaron used briefly.}
\usage{
group_by_shifted_day(
  df,
  hr_delay = 0,
  drop_group_posix = TRUE,
  group_output = TRUE,
  uday = FALSE
)
}
\arguments{
\item{df}{is a data frame with columns: SubjectID, posix, and activity}

\item{hr_delay}{time since midnight when a "new day" starts. For example, for max10 hours, we may wish day-groups to start at 5am (hr_delay = 5) rather than midnight. Alternatively, for total sleep time, we may wish days to start at 6pm (hr_delay = 18) when computing total sleep time.}
}
\value{
a data frame with: SubjectID, group_year, group_yday, and [summary measure]
}
\description{
Not maintained. (newer code should use Date rather than year + yday.
}
