% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_processing.r
\name{plot_sorted_viterbi}
\alias{plot_sorted_viterbi}
\title{plot results from get_viterbi_with_columns_sorted_by_state}
\usage{
plot_sorted_viterbi(
  vit,
  y,
  time_bin,
  index_to_plot = 1:length(y),
  groups = rep(1, length(y)),
  ylab = "probability",
  xlab = "time bin",
  means_included_in_level_names = FALSE,
  filllab = paste0("states", ifelse(means_included_in_level_names,
    " (within-\\nstate mean)", ""))
)
}
\description{
plot results from get_viterbi_with_columns_sorted_by_state
}
