% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAccel.R
\name{readAccel}
\alias{readAccel}
\title{Read in accelerometer file}
\usage{
readAccel(x)
}
\arguments{
\item{x}{Location of accelerometer csv file}
}
\description{
This function reads in an accelerometer file created by Oregon Research Electronics AL100 Acceleration Loggers. 
Returns a dataframe with 4 columns: sample number, and 3 acceleration axes
}
\examples{
readAccel()
}
