% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_data.R
\docType{data}
\name{ref_data}
\alias{ref_data}
\title{Reference data for Accelerometer testing}
\format{
An object of class \code{"tibble"}
}
\usage{
data(ref_data)
}
\description{
Reference data for Accelerometer testing
}
\examples{

ref_data[1:5,1:6]
}
\references{
https://github.com/activityMonitoring/biobankAccelerometerAnalysis
}
\keyword{datasets}
