% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmOut.R
\name{rmOut}
\alias{rmOut}
\title{Remove outliers from a dominant period series}
\usage{
rmOut(dat)
}
\arguments{
\item{dat}{A 'long form' data frame with three columns: 'day', 'value', and 'weights'. Value is the dominant period.}
}
\description{
This function removes outliers from a dominant period series, based residuals of a loess curve.  Returns the same data frame, but with NA's in rows of outliers.
}
\examples{
rmOut()
}
