% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_time.r
\name{shift_forward_by}
\alias{shift_forward_by}
\title{Shift time series forward, wrapping around (optionally with multiple channels)}
\usage{
shift_forward_by(x, i, L = dim(x)[1])
}
\arguments{
\item{x}{a matrix with column per signal channel}
}
\description{
Shift time series forward, wrapping around (optionally with multiple channels)
}
